+++ package.json
@@ -36,6 +36,7 @@
   "devDependencies": {
     "@types/jest": "^23.3.10",
     "@types/node": "^10.7.0",
+    "@types/request": "^2.48.1",
     "gulp": "^4.0.0",
     "gulp-rename": "^1.4.0",
     "gulp-sourcemaps": "^2.6.4",
@@ -44,6 +45,7 @@
     "gulp-uglify": "^3.0.1",
     "jest": "^23.6.0",
     "jest-junit": "^5.1.0",
+    "request": "^2.88.0",
     "semver": "^5.6.0",
     "source-map-loader": "^0.2.3",
     "ts-jest": "^23.10.5",
