+++ tests/TranslationRecognizerTests.ts
@@ -632,7 +632,7 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
             });
     }, 10000);
 
-    test.skip("MultiPhrase", (done: jest.DoneCallback) => {
+    test("MultiPhrase", (done: jest.DoneCallback) => {
         // tslint:disable-next-line:no-console
         console.info("Name: MultiPhrase");
         const s: sdk.SpeechTranslationConfig = BuildSpeechConfig();
@@ -646,9 +646,11 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
         const p: sdk.PushAudioInputStream = sdk.AudioInputStream.createPushStream();
         const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
         const numPhrases: number = 3;
+        const silentBuffer: ArrayBuffer = new ArrayBuffer(32000);
 
         for (let i: number = 0; i < 3; i++) {
             p.write(f);
+            p.write(silentBuffer);
         }
 
         p.close();
@@ -686,18 +688,13 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
         let inTurn: boolean = false;
 
         r.canceled = ((o: sdk.Recognizer, e: sdk.TranslationRecognitionCanceledEventArgs) => {
-            try {
-                switch (e.reason) {
-                    case sdk.CancellationReason.Error:
-                        done.fail(e.errorDetails);
-                        break;
-                    case sdk.CancellationReason.EndOfStream:
-                        expect(synthCount).toEqual(numPhrases);
-                        canceled = true;
-                        break;
-                }
-            } catch (error) {
-                done.fail(error);
+            switch (e.reason) {
+                case sdk.CancellationReason.Error:
+                    done.fail(e.errorDetails);
+                    break;
+                case sdk.CancellationReason.EndOfStream:
+                    canceled = true;
+                    break;
             }
         });
 
@@ -718,6 +715,7 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
 
                     for (let i: number = 0; i < synthFragmentCount; i++) {
                         p.write(rEvents[i]);
+                        p.write(silentBuffer);
                     }
                     p.close();
 
@@ -764,8 +762,14 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
                         WaitForCondition(() => (canceled && !inTurn),
                             () => {
                                 r2.stopContinuousRecognitionAsync(() => {
-                                    expect(numEvents).toEqual(numPhrases);
-                                    done();
+                                    try {
+                                        expect(synthCount).toEqual(numPhrases);
+                                        expect(numEvents).toEqual(numPhrases);
+                                        done();
+                                    } catch (error) {
+                                        done.fail(error);
+                                    }
+
                                 }, (error: string) => {
                                     done.fail(error);
                                 });
@@ -1245,152 +1249,6 @@ describe.each([true, false])("Service based tests", (forceNodeWebSocket: boolean
                 done.fail(err);
             });
     }, 35000);
-
-    // Tests client reconnect after speech timeouts.
-    test.skip("Reconnect After timeout", (done: jest.DoneCallback) => {
-        // tslint:disable-next-line:no-console
-        console.info("Name: Reconnect After timeout");
-        // Pump valid speech and then silence until at least one speech end cycle hits.
-        const fileBuffer: ArrayBuffer = WaveFileAudioInput.LoadArrayFromFile(Settings.WaveFile);
-
-        const alternatePhraseFileBuffer: ArrayBuffer = WaveFileAudioInput.LoadArrayFromFile(Settings.LuisWaveFile);
-
-        let p: sdk.PullAudioInputStream;
-        let s: sdk.SpeechTranslationConfig;
-        if (undefined === Settings.SpeechTimeoutEndpoint || undefined === Settings.SpeechTimeoutKey) {
-            // tslint:disable-next-line:no-console
-            console.warn("Running timeout test against production, this will be very slow...");
-            s = BuildSpeechConfig();
-        } else {
-            s = sdk.SpeechTranslationConfig.fromEndpoint(new URL(Settings.SpeechTimeoutEndpoint), Settings.SpeechTimeoutKey);
-        }
-        objsToClose.push(s);
-
-        s.addTargetLanguage(Settings.WaveFileLanguage);
-        s.speechRecognitionLanguage = Settings.WaveFileLanguage;
-
-        let pumpSilence: boolean = false;
-        let sendAlternateFile: boolean = false;
-
-        let bytesSent: number = 0;
-        const targetLoops: number = 250;
-
-        // Pump the audio from the wave file specified with 1 second silence between iterations indefinetly.
-        p = sdk.AudioInputStream.createPullStream(
-            {
-                close: () => { return; },
-                read: (buffer: ArrayBuffer): number => {
-                    if (pumpSilence) {
-                        bytesSent += buffer.byteLength;
-                        if (bytesSent >= 16000) {
-                            bytesSent = 0;
-                            pumpSilence = false;
-                        }
-                        return buffer.byteLength;
-                    } else {
-                        // Alternate between the two files with different phrases in them.
-                        const sendBuffer: ArrayBuffer = sendAlternateFile ? alternatePhraseFileBuffer : fileBuffer;
-
-                        const copyArray: Uint8Array = new Uint8Array(buffer);
-                        const start: number = bytesSent;
-                        const end: number = buffer.byteLength > (sendBuffer.byteLength - bytesSent) ? (sendBuffer.byteLength - 1) : (bytesSent + buffer.byteLength - 1);
-                        copyArray.set(new Uint8Array(sendBuffer.slice(start, end)));
-                        const readyToSend: number = (end - start) + 1;
-                        bytesSent += readyToSend;
-
-                        if (readyToSend < buffer.byteLength) {
-                            bytesSent = 0;
-                            pumpSilence = true;
-                            sendAlternateFile = !sendAlternateFile;
-                        }
-
-                        return readyToSend;
-                    }
-
-                },
-            });
-
-        const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-
-        const r: sdk.TranslationRecognizer = new sdk.TranslationRecognizer(s, config);
-        objsToClose.push(r);
-
-        let speechEnded: number = 0;
-        let lastOffset: number = 0;
-        let recogCount: number = 0;
-        let canceled: boolean = false;
-        let inTurn: boolean = false;
-        let alternatePhrase: boolean = false;
-
-        r.recognized = (o: sdk.Recognizer, e: sdk.TranslationRecognitionEventArgs) => {
-            try {
-                // If the target number of loops has been seen already, don't check as the audio being sent could have been clipped randomly during a phrase,
-                // and failing because of that isn't warranted.
-                if (recogCount <= targetLoops) {
-                    expect(sdk.ResultReason[e.result.reason]).toEqual(sdk.ResultReason[sdk.ResultReason.TranslatedSpeech]);
-                    expect(e.offset).toBeGreaterThanOrEqual(lastOffset);
-                    lastOffset = e.offset;
-
-                    // If there is silence exactly at the moment of disconnect, an extra speech.phrase with text ="" is returned just before the
-                    // connection is disconnected.
-                    if ("" !== e.result.text) {
-                        if (alternatePhrase) {
-                            expect(e.result.text).toEqual(Settings.LuisWavFileText);
-                        } else {
-                            expect(e.result.text).toEqual(Settings.WaveFileText);
-                        }
-
-                        alternatePhrase = !alternatePhrase;
-                    }
-                    if (recogCount++ >= targetLoops) {
-                        p.close();
-                    }
-                }
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.canceled = (o: sdk.Recognizer, e: sdk.TranslationRecognitionCanceledEventArgs): void => {
-            try {
-                expect(e.errorDetails).toBeUndefined();
-                expect(sdk.CancellationReason[e.reason]).toEqual(sdk.CancellationReason[sdk.CancellationReason.EndOfStream]);
-                canceled = true;
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.sessionStarted = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = true;
-        });
-
-        r.sessionStopped = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = false;
-        });
-
-        r.speechEndDetected = (o: sdk.Recognizer, e: sdk.RecognitionEventArgs): void => {
-            speechEnded++;
-        };
-
-        r.startContinuousRecognitionAsync(() => {
-            WaitForCondition(() => (canceled && !inTurn), () => {
-                r.stopContinuousRecognitionAsync(() => {
-                    try {
-                        expect(speechEnded).toEqual(1);
-                        done();
-                    } catch (error) {
-                        done.fail(error);
-                    }
-                }, (error: string) => {
-                    done.fail(error);
-                });
-            });
-        },
-            (err: string) => {
-                done.fail(err);
-            });
-    }, 1000 * 60 * 12);
 
     test("Silence Then Speech", (done: jest.DoneCallback) => {
         // tslint:disable-next-line:no-console
